[PaymentRequest] Move providers to PaymentBundle
