+++ src/Sylius/Bundle/ApiBundle/CommandHandler/Payment/AddPaymentRequestHandler.php
@@ -16,8 +16,8 @@ namespace Sylius\Bundle\ApiBundle\CommandHandler\Payment;
 use Sylius\Bundle\ApiBundle\Command\Payment\AddPaymentRequest;
 use Sylius\Bundle\ApiBundle\Exception\PaymentMethodNotFoundException;
 use Sylius\Bundle\ApiBundle\Exception\PaymentNotFoundException;
-use Sylius\Bundle\CoreBundle\PaymentRequest\Provider\DefaultActionProviderInterface;
-use Sylius\Bundle\CoreBundle\PaymentRequest\Provider\DefaultPayloadProviderInterface;
+use Sylius\Bundle\PaymentBundle\Provider\DefaultActionProviderInterface;
+use Sylius\Bundle\PaymentBundle\Provider\DefaultPayloadProviderInterface;
 use Sylius\Component\Core\Model\PaymentInterface;
 use Sylius\Component\Core\Model\PaymentMethodInterface;
 use Sylius\Component\Core\Repository\PaymentMethodRepositoryInterface;
