+++ src/Sylius/Bundle/PaymentBundle/Provider/DefaultActionProvider.php
@@ -11,9 +11,9 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\CoreBundle\PaymentRequest\Provider;
+namespace Sylius\Bundle\PaymentBundle\Provider;
 
-use Sylius\Bundle\PayumBundle\Model\GatewayConfigInterface;
+use Sylius\Component\Payment\Model\GatewayConfigInterface;
 use Sylius\Component\Payment\Model\PaymentRequestInterface;
 
 final class DefaultActionProvider implements DefaultActionProviderInterface
