+++ src/Sylius/Bundle/PaymentBundle/Provider/DefaultPayloadProvider.php
@@ -11,11 +11,14 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\CoreBundle\PaymentRequest\Provider;
+namespace Sylius\Bundle\PaymentBundle\Provider;
 
 use Sylius\Component\Payment\Model\PaymentRequestInterface;
 
-interface DefaultActionProviderInterface
+final class DefaultPayloadProvider implements DefaultPayloadProviderInterface
 {
-    public function getAction(PaymentRequestInterface $paymentRequest): string;
+    public function getPayload(PaymentRequestInterface $paymentRequest): mixed
+    {
+        return null;
+    }
 }
