+++ src/Sylius/Bundle/PaymentBundle/Provider/ServiceProviderAwareProviderInterface.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\CoreBundle\PaymentRequest\Provider;
+namespace Sylius\Bundle\PaymentBundle\Provider;
 
 interface ServiceProviderAwareProviderInterface extends HttpResponseProviderInterface
 {
