+++ src/Sylius/Bundle/PaymentBundle/Resources/config/services/command_provider.xml
@@ -21,7 +21,7 @@
 
         <service id="sylius.payment_request.command_provider.gateway_factory" class="Sylius\Bundle\PaymentBundle\CommandProvider\GatewayFactoryCommandProvider">
             <argument type="service" id="sylius.checker.payment_request_duplication" />
-            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Provider\GatewayFactoryNameProviderInterface" />
+            <argument type="service" id="sylius.payment_request.provider.gateway_factory_name" />
             <argument type="tagged_locator" tag="sylius.payment_request.command_provider" index-by="gateway-factory" />
         </service>
         <service id="sylius.payment_request.command_provider.default" alias="sylius.payment_request.command_provider.gateway_factory" />
