+++ src/Sylius/Bundle/PayumBundle/Resources/config/services/payment_request/command_handler.xml
@@ -20,7 +20,7 @@
         <defaults public="true" />
 
         <service id="sylius.payum.payment_request.command_handler.token_payment_request" class="Sylius\Bundle\PayumBundle\PaymentRequest\CommandHandler\TokenPaymentRequestHandler" abstract="true">
-            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Provider\PaymentRequestProviderInterface" />
+            <argument type="service" id="sylius.payment_request.provider.payment_request" />
             <argument type="service" id="sylius.payum.payment_request.resolver.doctrine_proxy_object_resolver" />
             <argument type="service" id="sylius.payum.payment_request.factory.payum_token" />
             <argument type="service" id="sylius.payum.payment_request.processor.request" />
@@ -38,7 +38,7 @@
         </service>
 
         <service id="sylius.payum.payment_request.command_handler.model_payment_request" class="Sylius\Bundle\PayumBundle\PaymentRequest\CommandHandler\ModelPaymentRequestHandler" abstract="true">
-            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Provider\PaymentRequestProviderInterface" />
+            <argument type="service" id="sylius.payment_request.provider.payment_request" />
             <argument type="service" id="sylius.payum.payment_request.resolver.doctrine_proxy_object_resolver" />
             <argument type="service" id="sylius.payum.payment_request.processor.request" />
         </service>
